gcp-bookstore-data-project/gcs-to-gbq-event-trigger-cloudfunc/requirements.txt
# cloudrun gcs-to-bigquery
functions-framework==3.*
google-cloud-storage
google-cloud-bigquery